generate

+++ v2/types/factories/IReceiverEVM.sol/IReceiverEVMEvents__factory.ts
@@ -144,8 +144,8 @@ const _abi = [
           },
           {
             name: "amount",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
           {
             name: "revertMessage",
